- imported latest shared enums - deleted duplicate local enums - renamed an enum file for consistency - declared  "empty restoration state" object - initialized store using "empty" object
