+++ package.json
@@ -14,25 +14,25 @@
   },
   "dependencies": {
     "@babel/compat-data": "^7.19.1",
-    "@bcrs-shared-components/action-chip": "1.1.3",
-    "@bcrs-shared-components/breadcrumb": "2.1.3",
-    "@bcrs-shared-components/business-lookup": "1.1.6",
-    "@bcrs-shared-components/certify": "2.1.3",
-    "@bcrs-shared-components/completing-party": "2.1.3",
+    "@bcrs-shared-components/action-chip": "1.1.5",
+    "@bcrs-shared-components/breadcrumb": "2.1.5",
+    "@bcrs-shared-components/business-lookup": "1.1.7",
+    "@bcrs-shared-components/certify": "2.1.5",
+    "@bcrs-shared-components/completing-party": "2.1.5",
     "@bcrs-shared-components/confirm-dialog": "1.2.1",
-    "@bcrs-shared-components/contact-info": "1.2.3",
+    "@bcrs-shared-components/contact-info": "1.2.5",
     "@bcrs-shared-components/corp-type-module": "1.0.8",
     "@bcrs-shared-components/court-order-poa": "2.1.3",
-    "@bcrs-shared-components/date-picker": "1.2.3",
+    "@bcrs-shared-components/date-picker": "1.2.5",
     "@bcrs-shared-components/detail-comment": "1.1.1",
-    "@bcrs-shared-components/enums": "1.0.25",
-    "@bcrs-shared-components/fee-summary": "1.2.3",
+    "@bcrs-shared-components/enums": "1.0.27",
+    "@bcrs-shared-components/fee-summary": "1.2.5",
     "@bcrs-shared-components/folio-number": "1.1.1",
     "@bcrs-shared-components/help-business-number": "1.1.1",
-    "@bcrs-shared-components/interfaces": "1.0.46",
-    "@bcrs-shared-components/nature-of-business": "1.2.2",
-    "@bcrs-shared-components/share-structure": "2.1.3",
-    "@bcrs-shared-components/staff-payment": "2.1.3",
+    "@bcrs-shared-components/interfaces": "1.0.48",
+    "@bcrs-shared-components/nature-of-business": "1.2.4",
+    "@bcrs-shared-components/share-structure": "2.1.4",
+    "@bcrs-shared-components/staff-payment": "2.1.4",
     "@mdi/font": "^5.9.55",
     "@sentry/browser": "^5.30.0",
     "@sentry/integrations": "^5.30.0",
