--- src/enums/filingNames.ts
@@ -1,13 +1,0 @@
-// *** TODO: move to shared components
-export enum FilingNames {
-  ALTERATION = 'Alteration',
-  CHANGE_OF_REGISTRATION = 'Change of Registration',
-  CONVERSION = 'Conversion',
-  CORRECTION = 'Correction',
-  DISSOLUTION_FIRM = 'Dissolution',
-  INCORPORATION_APPLICATION = 'Incorporation Application',
-  REGISTRATION = 'Registration',
-  RESTORATION_APPLICATION = 'Restoration Application',
-  SPECIAL_RESOLUTION = 'Special Resolution',
-  VOLUNTARY_DISSOLUTION = 'Voluntary Dissolution',
-}
