--- src/enums/filingTypes.ts
@@ -1,13 +1,0 @@
-// *** TODO: move to shared components
-export enum FilingTypes {
-  ALTERATION = 'alteration',
-  CHANGE_OF_REGISTRATION = 'changeOfRegistration',
-  CONVERSION = 'conversion',
-  CORRECTION = 'correction',
-  DISSOLUTION = 'dissolution',
-  INCORPORATION_APPLICATION = 'incorporationApplication',
-  REGISTRATION = 'registration',
-  RESTORATION = 'restoration',
-  SPECIAL_RESOLUTION = 'specialResolution',
-  VOLUNTARY_DISSOLUTION = 'dissolution',
-}
