+++ src/enums/index.ts
@@ -7,18 +7,15 @@ export * from './correctionErrorTypes'
 export * from './effectiveDateTypes'
 export * from './effectOfOrders'
 export * from './entitySubtypes'
-export * from './filingCodes'
-export * from './filingNames'
 export * from './filingStatus'
-export * from './filingTypes'
 export * from './partyTypes'
-export * from './NameChangeOptions'
+export * from './nameChangeOptions'
 export * from './nameRequestStates'
 export * from './nameRequestTypes'
-export * from './restorationTypes'
 export * from './roleTypes'
 export * from './routeNames'
 
 // external enum
 export { CorpTypeCd } from '@bcrs-shared-components/corp-type-module/'
-export { ActionTypes, EntityStates } from '@bcrs-shared-components/enums'
+export { ActionTypes, EntityStates, FilingCodes, FilingNames, FilingTypes, RestorationTypes }
+  from '@bcrs-shared-components/enums'
