--- src/enums/restorationTypes.ts
@@ -1,6 +1,0 @@
-export enum RestorationTypes {
-  FULL = 'fullRestoration',
-  LIMITED = 'limitedRestoration',
-  LTD_EXTEND = 'limitedRestorationExtension',
-  LTD_TO_FULL = 'limitedRestorationToFull'
-}
